+++ aiobungie/internal/enums.py
@@ -72,6 +72,9 @@ class Enum(__enum.Enum):
     def __str__(self) -> str:
         return self.name
 
+    def __repr__(self) -> str:
+        return self.__str__()
+
     def __int__(self) -> int:
         return int(self.value)